Refact to MongoDB

+++ src/main/java/learningspringboot/TeamRepository.java
@@ -1,3 +1,6 @@
 package learningspringboot;
+import java.math.BigInteger;
+
 import org.springframework.data.repository.CrudRepository;
-public interface TeamRepository extends CrudRepository<Team, Long> {}
+public interface TeamRepository extends CrudRepository<Team, BigInteger> 
+{}
