+++ src/main/java/learningspringboot/TeammateRepository.java
@@ -1,4 +1,6 @@
 package learningspringboot;
+import java.math.BigInteger;
+
 import org.springframework.data.repository.CrudRepository;
-public interface TeammateRepository extends CrudRepository<Teammate, Long>
+public interface TeammateRepository extends CrudRepository<Teammate, BigInteger>
 {}
